Incroporating Piotr's feedback.
- disabled by default (decided not to introduce env variable)
- inter-connecting `AudioCache` with `caching.py`, so `set_caching_enabled()` activates/deactivates AudioCache
- cache is freed upon disabling
- removed @dynamic_lru_cache from `audio.py` (in `features/io.py` it is not removed)

The code of `AudioCache` code was moved to `caching.py`.

And the new unit tests were created.
Should I also remove the old unit tests?
`test_audio_caching_disabled_works`
`test_audio_caching_enabled_works`

Cheers,
Karel
