+++ lhotse/caching.py
@@ -1,5 +1,7 @@
+import os
 from functools import lru_cache, wraps
-from typing import Any, Callable
+from threading import Lock
+from typing import Any, Callable, Dict, Union
 
 LHOTSE_CACHING_ENABLED = False
 
@@ -16,6 +18,10 @@ def set_caching_enabled(enabled: bool) -> None:
     global LHOTSE_CACHED_METHOD_REGISTRY
     assert isinstance(enabled, bool)
     LHOTSE_CACHING_ENABLED = enabled
+
+    # enable cache for audio files of "command" type
+    AudioCache.enable(enabled)
+
     if not enabled:
         # Caching disabled: purge all caches.
         for method in LHOTSE_CACHED_METHOD_REGISTRY["cached"].values():
@@ -70,3 +76,104 @@ def dynamic_lru_cache(method: Callable) -> Callable:
         return m(*args, **kwargs)
 
     return wrapper
+
+
+class AudioCache:
+    """
+    Cache of 'bytes' objects with audio data.
+    It is used to cache the "command" type audio inputs.
+
+    By default it is disabled, to enable call `set_caching_enabled(True)`
+    or `AudioCache.enable()`.
+
+    The cache size is limited to max 100 elements and 500MB of audio.
+
+    A global dict `__cache_dict` (static member variable of class AudioCache)
+    is holding the wavs as 'bytes' arrays.
+    The key is the 'source' identifier (i.e. the command for loading the data).
+
+    Thread-safety is ensured by a threading.Lock guard.
+    """
+
+    __enabled: bool = False
+
+    max_cache_memory: int = 500 * 1e6  # 500 MB
+    max_cache_elements: int = 100  # 100 audio files
+
+    __cache_dict: Dict[str, bytes] = {}
+    __lock: Lock = Lock()
+
+    @classmethod
+    def enable(cls, enabled=True):
+        cls.__enabled = enabled
+        if not enabled:
+            cls.__clear_cache()
+
+    @classmethod
+    def enabled(cls) -> bool:
+        return cls.__enabled
+
+    @classmethod
+    def try_cache(cls, key: str) -> Union[bytes, None]:
+        """
+        Test if 'key' is in the chache. If yes return the bytes array,
+        otherwise return None.
+        """
+
+        if not cls.__enabled:
+            return None
+
+        with cls.__lock:
+            if key in cls.__cache_dict:
+                return cls.__cache_dict[key]
+            else:
+                return None
+
+    @classmethod
+    def add_to_cache(cls, key: str, value: bytes):
+        """
+        Add the new (key,value) pair to cache.
+        Possibly free some elements before adding the new pair.
+        The oldest elements are removed first.
+        """
+
+        if not cls.__enabled:
+            return None
+
+        if len(value) > cls.max_cache_memory:
+            return
+
+        with cls.__lock:
+            # limit cache elements
+            while len(cls.__cache_dict) > cls.max_cache_elements:
+                # remove oldest elements from cache
+                # (dict pairs are sorted according to insertion order)
+                cls.__cache_dict.pop(next(iter(cls.__cache_dict)))
+
+            # limit cache memory
+            while len(value) + AudioCache.__cache_memory() > cls.max_cache_memory:
+                # remove oldest elements from cache
+                # (dict pairs are sorted according to insertion order)
+                cls.__cache_dict.pop(next(iter(cls.__cache_dict)))
+
+            # store the new (key,value) pair
+            cls.__cache_dict[key] = value
+
+    @classmethod
+    def __cache_memory(cls) -> int:
+        """
+        Return size of AudioCache values in bytes.
+        (internal, not to be called from outside)
+        """
+        ans = 0
+        for key, value in cls.__cache_dict.items():
+            ans += len(value)
+        return ans
+
+    @classmethod
+    def __clear_cache(cls) -> None:
+        """
+        Clear the cache, remove the data.
+        """
+        with cls.__lock:
+            cls.__cache_dict.clear()
